+++ BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/exceptions/AppExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.github.coding_team_sept.nd_backend.appointment.exceptions;
 
-import com.github.coding_team_sept.nd_backend.appointment.payload.responses.ErrorResponse;
+import com.github.coding_team_sept.nd_backend.appointment.payloads.responses.ErrorResponse;
 import org.junit.jupiter.api.Order;
 import org.springframework.core.Ordered;
 import org.springframework.http.HttpStatus;
