--- BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payload/responses/AppResponse.java
@@ -1,4 +1,0 @@
-package com.github.coding_team_sept.nd_backend.appointment.payload.responses;
-
-public abstract class AppResponse {
-}
