+++ BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payloads/requests/AppointmentRequest.java
@@ -1,4 +1,4 @@
-package com.github.coding_team_sept.nd_backend.appointment.payload.requests;
+package com.github.coding_team_sept.nd_backend.appointment.payloads.requests;
 
 public record AppointmentRequest(
         Long doctorId,
