+++ BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payloads/responses/AppResponse.java
@@ -1,0 +1,4 @@
+package com.github.coding_team_sept.nd_backend.appointment.payloads.responses;
+
+public abstract class AppResponse {
+}
