+++ BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payloads/responses/DataResponse.java
@@ -1,4 +1,4 @@
-package com.github.coding_team_sept.nd_backend.appointment.payload.responses;
+package com.github.coding_team_sept.nd_backend.appointment.payloads.responses;
 
 public class DataResponse extends AppResponse {
     public DataResponse() {}
