+++ BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payloads/responses/ResponseWrapper.java
@@ -1,4 +1,4 @@
-package com.github.coding_team_sept.nd_backend.appointment.payload.responses;
+package com.github.coding_team_sept.nd_backend.appointment.payloads.responses;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
