+++ BackEnd/appointment/src/main/java/com/github/coding_team_sept/nd_backend/appointment/payloads/responses/UsersDataResponse.java
@@ -1,4 +1,4 @@
-package com.github.coding_team_sept.nd_backend.appointment.payload.responses;
+package com.github.coding_team_sept.nd_backend.appointment.payloads.responses;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
